user_authentication_system_and_api_security_validation/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt
slowapi
passlib
email-validator 